+++ app/javascript/vue/components/radials/annotator/annotator.vue
@@ -4,7 +4,7 @@
       class="radial-annotator">
       <modal
         v-if="display"
-        :container-style="{ backgroundColor: 'transparent' }"
+        :container-style="{ backgroundColor: 'transparent', boxShadow: 'none' }"
         @close="closeModal()">
         <h3
           slot="header"